Remove special handling of OS X
On newer versions of macOS `systemsetup -gettimezone` required sudo privileges
and hence was never used. Instead the fallback of looking at where /etc/localtime
pointed was used, and this is a part of the generic unix support as well.
So we remove all special darwin handling and use the unix way.